continue worning on post functionality
